[Multi Apps] Change remove call to take list
Currently, the add call takes in a list of apps to install but the
remove call only takes a single app id path. This change matches the
remove call to the add call to also take in a list. Otherwise, the
developer would have to do an API call for each app they want to
uninstall, which could be many at once.

Bug: 1382072
Change-Id: I458b7a35bebbb04d541a8ea7872ad63a6de28bb2