[@scope] Make '&' in <scope-end> resolve to <scope-start>
Instead of storing the <scope-start> (i.e. From()) as a CSSSelectorList,
store an empty StyleRule with that same list of selectors.

This allows us to pretend that <scope-start> is the ancestor rule
for nesting purposes when matching <scope-end>, as required by spec.

The nesting selector (&) needs to work other places too, but that is
for future CLs to handle.

https://drafts.csswg.org/css-nesting-1/#nesting-at-scope

Bug: 1280240
Change-Id: I52bed7537a6892746bbf638407839abbd11ba96f