Consistency check between AXObjectCacheImpl and AXTreeSerializer tree
This is the lite version of checking the entire tree structure, only checking that the number of nodes is the same. It's easier to do and a more performant check. This CL only adds the checks. Four precursor CLs were landed to clean up object leaks and tree inconsistencies so that these checks would pass: AXMenuList (CL:4289642), Slots (CL:4294848), Image maps (CL:4291697), Inline text boxes (CL:4289593).

Follow up: a future check should be added to check for consistency between AXTreeSerializer and AXTree.

Bug: none
Change-Id: I0f33f9f7bdcd3f70cfce9182ab1111c7dfec6236
Cq-Do-Not-Cancel-Tryjobs: true