Add :visited duplicate rules to regular buckets
During RuleSet::AddRule, when adding a rule which depends on visited
link status, we add a duplicate rule (but with different a
LinkMatchType) to a special bucket for visited rules. However, this
approach does not quite work, since we depend on certain rules being
present in specific buckets in some situations. In particular, rules
which contain ::part() must be in the part bucket, because we don't
provide the relevant part names when when matching rules from other
buckets.

This CL drops the special 'visited_dependent_rules_' and instead
adds the duplicate rules via FindBestRuleSetAndAdd.

Note that the function StyleResolver::StyleRulesForElement (only
used for Editing) now needs to suppress :visited styles (see
SerializationTest.Link and crbug.com/938590), because we try to match
the visited-dependent rules regardless of the value of
ElementRuleCollector::inside_link_.

Fixed: 1413904
Change-Id: Id360bef4f1b977107f109c4520db08f6e602a034